Add footer, adjust heading link placement
Added footer to the page, added icons for the apps link and the sign in link.  Removed border from searchbox so that it just uses the box-shadow to create a border, which is thinner than the minimum setting for a true border.  Added padding to top of div containing searchbox and buttons to mimic the google homepage.